@@ -5,7 +5,7 @@ homepage: https://www.shield.com
 
 environment:
   sdk: ">=2.12.0 <4.0.0"
-  flutter: ">=1.20.0"
+  flutter: ">=3.10.0"
 
 dependencies:
   flutter:
@@ -15,7 +15,7 @@ dependencies:
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_lints: ^1.0.0
+flutter_lints: ^4.0.0
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
