@@ -7,24 +7,17 @@ import 'shield_config.dart';
 class Shield {
   static const MethodChannel _channel = MethodChannel('plugin_shieldfraud');
   static ShieldCallback? _shieldCallback;
+
   static initShield(ShieldConfig config) async {
     try {
       if (config.shieldCallback != null) {
         _channel.setMethodCallHandler(_methodHandler);
         _shieldCallback = config.shieldCallback;
       }
-      String environment = "prod";
-      switch (config.environment) {
-        case ShieldEnvironment.dev:
-          environment = "dev";
-          break;
-        case ShieldEnvironment.prod:
-          environment = "prod";
-          break;
-        case ShieldEnvironment.staging:
-          environment = "staging";
-          break;
-      }
+
+      // ✅ Send enum NAME directly (important)
+      String environment = config.environment.name;
+      String logLevel = config.logLevel.name;
 
       setCrossPlatformParameters();
 
@@ -36,18 +29,15 @@ class Shield {
         "enableMocking": config.enableMocking,
         "partnerId": config.partnerId,
         "environment": environment,
+        "logLevel": logLevel,
         "defaultBlockedDialog": config.defaultBlockedDialog != null
             ? {
-                "title": config.defaultBlockedDialog!.title,
-                "body": config.defaultBlockedDialog!.body,
-              }
+          "title": config.defaultBlockedDialog!.title,
+          "body": config.defaultBlockedDialog!.body,
+        }
             : null,
-        "logLevel": config.logLevel.toString()
       });
-    } catch (_) {
-      //something went wrong during initialization. nothing we can do
-    }
-    return;
+    } catch (_) {}
   }
 
   static Future<void> setCrossPlatformParameters() async {
@@ -58,83 +48,83 @@ class Shield {
   }
 
   static Future<String> get sessionId async {
-    String sessionID = "";
     try {
-      sessionID = await _channel.invokeMethod('getSessionID');
-    } catch (_) {}
-    return sessionID;
+      return await _channel.invokeMethod('getSessionID');
+    } catch (_) {
+      return "";
+    }
   }
 
   static ShieldError? latestError;
+
   static Future<Map<String, dynamic>?> get latestDeviceResult async {
-    var result = "";
     try {
-      result = await _channel.invokeMethod('getDeviceResult');
+      final result = await _channel.invokeMethod('getDeviceResult');
       latestError = null;
       return json.decode(result);
     } on PlatformException catch (e) {
       latestError =
-          ShieldError(int.parse(e.code), e.message ?? "Unknown error");
+          ShieldError(int.tryParse(e.code) ?? 0, e.message ?? "Unknown error");
     } catch (_) {
       latestError = ShieldError(0, "Unknown error");
     }
     return null;
   }
 
-  static Future<bool> sendAttributes(
-      String screenName, Map<String, String> data) async {
-    var status = false;
+  static Future<String?> sendAttributes(
+      String screenName,
+      Map<String, String> data,
+      ) async {
     try {
-      status = await _channel.invokeMethod(
-          "sendAttributes", {"screenName": screenName, "attributes": data});
-    } catch (_) {}
-    return status;
+      final result = await _channel.invokeMethod(
+        "sendAttributes",
+        {
+          "screenName": screenName,
+          "attributes": data,
+        },
+      );
+
+      return result as String?;
+    } catch (e) {
+      return null;
+    }
   }
 
+
   static Future<bool> sendDeviceSignature(String screenName) async {
-    var status = false;
     try {
-      status = await _channel.invokeMethod("sendDeviceSignature", {
+      return await _channel.invokeMethod("sendDeviceSignature", {
         "screenName": screenName
-      }).timeout(const Duration(seconds: 5), onTimeout: () {
-        //If the result not comeback in 5 seconds, just time out
-        return false;
-      });
-    } catch (_) {}
-    return status;
+      }).timeout(const Duration(seconds: 30), onTimeout: () => false);
+    } catch (_) {
+      return false;
+    }
   }
 
   static Future<bool> get isShieldInitialized async {
-    var isShieldInitialized = false;
     try {
-      isShieldInitialized = await _channel.invokeMethod("isShieldInitialized");
-    } catch (_) {}
-    return isShieldInitialized;
+      return await _channel.invokeMethod("isShieldInitialized");
+    } catch (_) {
+      return false;
+    }
   }
 
   static Future<void> _methodHandler(MethodCall call) async {
     try {
       switch (call.method) {
         case "setDeviceResult":
-          {
-            _shieldCallback?.onSuccess(json.decode(call.arguments));
-            break;
-          }
+          _shieldCallback?.onSuccess(json.decode(call.arguments));
+          break;
+
         case "setDeviceResultError":
-          {
-            ShieldError shieldError = ShieldError(call.arguments["code"] ?? 0,
-                call.arguments["message"] ?? "Unknown error");
-            _shieldCallback?.onError(shieldError);
-            break;
-          }
+          ShieldError shieldError = ShieldError(
+              call.arguments["code"] ?? 0,
+              call.arguments["message"] ?? "Unknown error");
+          _shieldCallback?.onError(shieldError);
+          break;
       }
-    } on PlatformException catch (e) {
-      ShieldError shieldError =
-          ShieldError(int.parse(e.code), e.message ?? "Unknown error");
-      _shieldCallback?.onError(shieldError);
     } catch (_) {
-      ShieldError shieldError = ShieldError(0, "Unknown error");
-      _shieldCallback?.onError(shieldError);
+      _shieldCallback?.onError(ShieldError(0, "Unknown error"));
     }
   }
 }
